+++ frigate/util/process.py
@@ -7,7 +7,10 @@ import threading
 from logging.handlers import QueueHandler
 from typing import Callable, Optional
 
+from setproctitle import setproctitle
+
 import frigate.log
+from frigate.config.logger import LoggerConfig
 
 
 class BaseProcess(mp.Process):
@@ -50,7 +53,9 @@ class Process(BaseProcess):
     def before_start(self) -> None:
         self.__log_queue = frigate.log.log_listener.queue
 
-    def pre_run_setup(self) -> None:
+    def pre_run_setup(self, logConfig: LoggerConfig | None = None) -> None:
+        setproctitle(self.name)
+        threading.current_thread().name = f"process:{self.name}"
         faulthandler.enable()
 
         def receiveSignal(signalNumber, frame):
@@ -68,3 +73,8 @@ class Process(BaseProcess):
         self.logger = logging.getLogger(self.name)
         logging.basicConfig(handlers=[], force=True)
         logging.getLogger().addHandler(QueueHandler(self.__log_queue))
+
+        if logConfig:
+            frigate.log.apply_log_levels(
+                logConfig.default.value.upper(), logConfig.logs
+            )
